+++ hw12_13_14_15_calendar/go.mod
@@ -1,3 +1,10 @@
-module github.com/fixme_my_friend/hw12_13_14_15_calendar
+module github.com/gt-edu/otus-golang-hw/hw12_13_14_15_calendar
 
-go 1.16
+go 1.18
+
+require github.com/spf13/cobra v1.6.1
+
+require (
+	github.com/inconshreveable/mousetrap v1.0.1 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)
